requirements.txt
unstructured
tiktoken
pinecone-client
pypdf
openai==0.28
langchain-community
pandas
numpy
python-dotenv
pinecone
flask
pymongo
-e .